Use CMake (instead of Makefile) for C extension
