--- interop/klr/.gitignore
@@ -1,2 +1,0 @@
-cbor_test
-simplify_test
