+++ interop/klr/frontend.h
@@ -4,21 +4,12 @@ Released under Apache 2.0 license as described in the file LICENSE.
 Authors: Paul Govereau, Sean McLaughlin
 */
 #pragma once
+
+// frontend.h (this file) is to support .c files that do NOT require the Python C API.
+// frontend_py.h is to support .c files that DO require the Python C API.
+
 #include "stdc.h"
 #include "region.h"
-
-#define PY_SSIZE_T_CLEAN
-#include <Python.h>
-static_assert(
-    PY_MAJOR_VERSION == 3 &&
-    PY_MINOR_VERSION >= 9 &&
-    PY_MINOR_VERSION <= 12,
-    "Unsupported Python Version");
-
-#if PY_MINOR_VERSION == 9
-#define Py_IsNone(x) ((x) == Py_None)
-#define Py_IsTrue(x) ((x) == Py_True)
-#endif
 
 // Front-end version (place holder)
 #define KLR_VERSION 1
@@ -27,26 +18,9 @@ static_assert(
 //#define MODULE_ROOT "neuronxcc.nki"
 #define MODULE_ROOT ""
 
-// The front-end is accessed through the class Kernel; one instance
-// per kernel. Each instance has a `struct kernel` on the C side.
-
-struct kernel {
-  PyObject_HEAD
-  PyObject *f;   // Kernel function
-  bool specialized;
-  struct region *python_region;
-  struct Python_Kernel *python_kernel;
-  struct region *nki_region;
-  struct NKI_Kernel *nki_kernel;
-};
-
-// peg_parser.c
-struct _mod* parse_string(const char *str, PyObject* filename);
-void free_python_ast(struct _mod *m);
-
-// gather.c
-bool gather(struct kernel *k);
-bool specialize(struct kernel *k, PyObject *args, PyObject *kws);
+// Forward declarations
+struct NKI_Kernel;
+struct Python_Kernel;
 
 // simplify.c
 struct SimpResult {
